🚚 Rename: icon → icons 폴더명 변경
