+++ src/components/search/add-info/result.tsx
@@ -2,7 +2,7 @@
 
 import Location from "@/assets/icons/location.svg";
 import Search from "@/assets/icons/search.svg";
-import Icon from "@/components/commons/icon";
+import Icon from "@/components/commons/icons";
 import { SectionLayout } from "@/components/layout";
 import { useSearchStore } from "@/stores/useSearchStore";
 import useUploadTextStore from "@/stores/useUploadTextStore";
